+++ src/Modules/Admin/Controllers/UserAuthController.php
@@ -63,7 +63,7 @@ class UserAuthController extends Controller
             ->allWithChildren();
 
         $provider = new ArrayDataProvider([
-            'allModels' => $items,
+            'allModels' => $this->module->sortAuthItems($items),
             'pagination' => false,
         ]);